Remove support for buckets and fix/improve timestamping
Bucket are no longer a supported Mixpanel feature so they have
been removed. Mixpanel has replaced this with segmentation.

Timestamps were being duplicated in the dictionary with different cases
so this has been fixed. As well there are more options for how to send
timestamps including not sending one (in which case Mixpanel sets the
time when it receives the event).  The default is to set it
automatically.